requirements.txt
Flask==1.0.2
PyJWT==1.6.4
